This PR has 2 parts: generational-GC and policy-improvements.

## I. Generational GC

Classify segments into 2 dimensions --- data category (data, metadata) and generation (0, 1, 2), totally 6 types of segments --- to improve the locality of block placements. This is the foundation to form a desired bimodal distribution of
segment utilizations for more efficient GC.

The profiling with rbd-fio (zipf-0.6) proves that the desired bimodal distribution has been formed. Note that reclaiming has been delayed as late as possible (see https://github.com/ceph/ceph/pull/46437), so that the first ~1500 seconds won't involve reclaiming, and the utility distribution below is accurate before 1500s:
* Before[^1]: all the blocks are mixed together so segment utilities are around 20~30%, which is not bimodal.
* After[^2]: nearly 55% of the segments have utilities below 10%, and another half have utilities around 55%, which is bimodal.

## II. Policy improvements

* `record_header_t` to store the average modify time instead of the commit time to consider dirty extents.
* Drop tracking rewrite-time, consider rewriting seems to have negative impacts: 0.55 WAF from reclaiming that considers rewrite-time or modify-time, and 0.33 WAF that only considers modify-time, in rbd-fio(zipf-0.6, 5400s).
* Drop the last-modify field in `extent_info_t`, which seems unnecessary right now.
* Make sure the information of modify-time won't be lost during rewriting.
* Introduce 3 GC policies to select from: greedy, benefit, and cost-benefit.

## Final performance improvements with rbd-fio, zipf-0.6, 5400s

* WAF from trimming is reduced from 2.03 to 1.32.
* WAF from reclaiming is reduced from 0.90 to 0.33.
* Overall conflict ratio are reduced, see before[^3] and after[^4].
* Overall throughput in the steady state is improved 11%.

## Policy selection

There are 3 GC policies available:
* GREEDY: always select the least utilized segment to reclaim.
* COST_BENEFIT: revise base on sprite-lfs's cost-benefit formula, and revise from the original version.
* BENEFIT: a revised version of GREEDY policy that considers the age factor, so that the aged segment has larger benefit:
![benefit-formula](https://user-images.githubusercontent.com/7736006/174752085-9ce7b6a6-8dc4-465c-8e66-d3a06ae64c69.png)

In theory, GREEDY policy should be the least efficient as it only considers segment utility. The result reclaim WAFs in rbd-fio(zipf-0.6, 5400s) are:
* GREEDY: 0.586
* COSY_BENEFIT: 0.328
* BENEFIT: 0.372

For reference, the original reclaim WAFs were:
* Before generational GC: 0.896
* After generational GC, before policy improvements: 0.553

[^1]: ![before-gen-dist](https://user-images.githubusercontent.com/7736006/174697816-e9705230-adb0-47e3-96a5-5d9a81df78a8.png)

[^2]: ![after-gen-dist](https://user-images.githubusercontent.com/7736006/174698322-9840c08b-628a-4ed8-b9d3-ccf711b2774e.png)

[^3]: ![before-gen-conflicts](https://user-images.githubusercontent.com/7736006/174703387-76125c93-d95c-4f0d-b37e-1e89c3292a19.png)

[^4]: ![after-gen-conflicts](https://user-images.githubusercontent.com/7736006/174703550-55f271a8-ad3c-4b00-814a-37ac63cbe966.png)

<!--
  - Please give your pull request a title like

      [component]: [short description]

  - Please use this format for each git commit message:

      [component]: [short description]

      [A longer multiline description]

      Fixes: [ticket URL on tracker.ceph.com, create one if necessary]
      Signed-off-by: [Your Name] <[your email]>

    For examples, use "git log".
-->

## Contribution Guidelines
- To sign and title your commits, please refer to [Submitting Patches to Ceph](https://github.com/ceph/ceph/blob/master/SubmittingPatches.rst).

- If you are submitting a fix for a stable branch (e.g. "pacific"), please refer to [Submitting Patches to Ceph - Backports](https://github.com/ceph/ceph/blob/master/SubmittingPatches-backports.rst) for the proper workflow.

## Checklist
- Tracker (select at least one)
  - [ ] References tracker ticket
  - [ ] Very recent bug; references commit where it was introduced
  - [ ] New feature (ticket optional)
  - [ ] Doc update (no ticket needed)
  - [ ] Code cleanup (no ticket needed)
- Component impact
  - [ ] Affects [Dashboard](https://tracker.ceph.com/projects/dashboard/issues/new), opened tracker ticket
  - [ ] Affects [Orchestrator](https://tracker.ceph.com/projects/orchestrator/issues/new), opened tracker ticket
  - [ ] No impact that needs to be tracked
- Documentation (select at least one)
  - [ ] Updates relevant documentation
  - [ ] No doc update is appropriate
- Tests (select at least one)
  - [ ] Includes [unit test(s)](https://docs.ceph.com/en/latest/dev/developer_guide/tests-unit-tests/)
  - [ ] Includes [integration test(s)](https://docs.ceph.com/en/latest/dev/developer_guide/testing_integration_tests/)
  - [ ] Includes bug reproducer
  - [ ] No tests

<details>
<summary>Show available Jenkins commands</summary>

- `jenkins retest this please`
- `jenkins test classic perf`
- `jenkins test crimson perf`
- `jenkins test signed`
- `jenkins test make check`
- `jenkins test make check arm64`
- `jenkins test submodules`
- `jenkins test dashboard`
- `jenkins test dashboard cephadm`
- `jenkins test api`
- `jenkins test docs`
- `jenkins render docs`
- `jenkins test ceph-volume all`
- `jenkins test ceph-volume tox`
- `jenkins test windows`
</details>