@xxhdx1985126 2022-06-22
> * After[2](#user-content-fn-2-4c532e593e0bd712a299837faa6a34fe): nearly 55% of the segments have utilities below 10%, and another half have utilities around 55%, which is bimodal.

It seems that, after the first 1500 secs, the number of segments whose utilization is under dropping, could you allow the test to keep running longer so that the stable segment usage distribution can emerge?

And could you elaborate a little bit about the test environment and the segment usage distribution graph? like how large the seastore disk space and the rbd image is, what the curves in the graph represent. It seems that the total number of segments in you seastore disk is about 5000~6000, so I guess you are not using the default seastore configuration, right?