@xxhdx1985126 2022-06-22
> so that the first ~1500 seconds won't involve reclaiming, and the utility distribution below is accurate before 1500s:

Do you mean bimodel distribution in the second graph is done by rewrites caused by journal trimming?