@cyx1231st 2022-06-22
> Do you mean that the bimodel distribution in the second graph is done by rewrites caused by journal trimming?

No, I mean metadata and data, or generation 0, 1, 2 have different probabilities/rates to become invalid. If we split write them into different segments, the ideal bimodel distribution can be formed automatically.