@cyx1231st 2022-06-22
> could you allow the test to keep running longer so that the stable segment usage distribution can emerge?

I wanted to show the bimodel utilility distribution before the involvements of reclaming, affected by the specific GC policy.

Anyway, this is the complete graph.
Before:
![before](https://user-images.githubusercontent.com/7736006/174928365-770b6faa-24d7-49f7-ae78-c655376485f7.png)

After:
![after](https://user-images.githubusercontent.com/7736006/174928374-db28b790-5fa7-4308-88ab-2d12c69f647d.png)