@cyx1231st 2022-06-22
> And could you elaborate a little bit about the test environment.

The disk and rbd image sizes are all 50G, the graph also shows the total number of segments is 800.

> the segment usage distribution graph?

See https://github.com/ceph/ceph/blob/3b983cccbb6465370081b5d798bc45a7e8d1bc15/src/crimson/os/seastore/segment_cleaner.h#L843-L866