@cyx1231st 2022-06-22
> Ah, so the segment usage distribution is showing the number of segments whose utilization is under a specific percentage, right?

Yes, and how they were evolved by time.